+++ app/components/oversee/resources/table.rb
@@ -9,7 +9,7 @@ class Oversee::Resources::Table < Oversee::Base
     @params = params
     @options = options
 
-    @oversee_resource = Oversee::Resource.new(resource_class: @resource_class)
+    @oversee_resource = Oversee::Resource.new(klass: @resource_class)
   end
 
   def view_template