+++ app/models/oversee/resource.rb
@@ -1,44 +1,53 @@
 class Oversee::Resource
   attr_reader :klass
-  attr_reader :resource_class
-  attr_reader :resource_class_name
-  attr_reader :instance
   attr_reader :associations
 
-  def initialize(klass: nil, resource_class: nil, instance: nil)
-    @klass = klass || resource_class
-    @resource_class = resource_class # For backwards compatibility
-
-    @instance = instance
+  def initialize(klass: nil)
+    @klass = klass
+    raise ArgumentError, "klass must be a class" unless klass.is_a?(Class)
   end
 
   # Route helpers
   def index_path
-    "/resources/#{resource_class_name}"
+    "/resources/#{class_name}"
   end
 
-  def show_path
-    # Rails.application.routes.url_helpers.resource_path(instance || object, resource_class_name:)
-    "/resources/#{resource_class_name}/#{instance.to_param}"
+  def show_path(record:)
+    # Rails.application.routes.url_helpers.resource_path(instance || object, class_name:)
+    "/resources/#{class_name}/#{record.to_param}"
   end
 
   # Columns
   def columns_for_create
-    excluded_columns = [resource_class.primary_key, "created_at", "updated_at"]
-    resource_class.columns_hash.except(*excluded_columns)
+    excluded_columns = [klass.primary_key, "created_at", "updated_at"]
+    klass.columns_hash.except(*excluded_columns)
   end
 
-  def columns_for_show
-
+  # Helpers
+  # Returns the class name of the entity.
+  #
+  # This method returns the name of the class as a string.
+  #
+  # @return [String] the name of the class
+  def class_name
+    @class_name ||= klass.to_s
   end
 
   # Associations
-  # Structured by association macro
+
+  # Returns the associations for the Entity.
+  #
+  # This method iterates over all associations of the resource class and maps them
+  # into a hash categorized by the association type (macro). Each association is
+  # represented as a hash containing details such as the name, class name, foreign key,
+  # whether it is optional, if it is a through association, and if it is a rich text association.
+  #
+  # @return [Hash] a hash where the keys are association types and the values are arrays of association details
   def associations
     @associations ||= begin
       map = Hash.new { |hash, key| hash[key] = [] }
 
-      @resource_class.reflect_on_all_associations.each do |association|
+      klass.reflect_on_all_associations.each do |association|
         map[association.macro] << {
           name: association.name,
           class_name: association.class_name,
@@ -52,17 +61,22 @@ class Oversee::Resource
     end
   end
 
+  # Returns the foreign keys for the Entity.
+  #
+  # This method iterates over the reflections of the resource class and collects
+  # the foreign keys for associations where the reflection belongs to another model.
+  #
+  # @return [Array<String>] an array of foreign key names
   def foreign_keys
-    resource_class.reflections.map do |name, reflection|
+    klass.reflections.map do |name, reflection|
       reflection.foreign_key if reflection.belongs_to?
     end.compact
   end
 
+  # Returns the rich text associations for the Entity.
+  #
+  # @return [Array<Hash>] an array of hashes representing the rich text associations
   def rich_text_associations
     @rich_text_associations ||= associations[:has_one].select { |association| association[:rich_text] }
-  end
-
-  def resource_class_name
-    @resource_class_name ||= resource_class.to_s
   end
 end